Rename package and update imports
